requirements.txt
absl-py==2.1.0
albucore==0.0.12
albumentations==1.4.10
annotated-types==0.7.0
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
grpcio==1.64.1
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
MedPy==0.5.1
mpmath==1.3.0
networkx==3.2.1
nibabel==5.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.2.0
protobuf==4.25.3
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
SimpleITK==2.3.1
six==1.16.0
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.6.18
tomli==2.0.1
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.9.0
Werkzeug==3.0.3
zipp==3.19.2
